+++ src/tool/components/tool-canvas.tsx
@@ -7,6 +7,7 @@ import {
   getRequiredRatio,
   formatRatio,
 } from "@/lib/contrast";
+import { CheckIcon, XIcon, PlusIcon } from "./icons";
 
 interface ToolCanvasProps {
   fgColor: string;
@@ -212,43 +213,7 @@ function ColorPreview({
   );
 }
 
-/** Check icon SVG */
-function CheckIcon() {
-  return (
-    <svg
-      xmlns="http://www.w3.org/2000/svg"
-      viewBox="0 0 20 20"
-      fill="currentColor"
-      style={{ width: "18px", height: "18px" }}
-      aria-hidden="true"
-    >
-      <path
-        fillRule="evenodd"
-        d="M16.707 5.293a1 1 0 010 1.414l-8 8a1 1 0 01-1.414 0l-4-4a1 1 0 011.414-1.414L8 12.586l7.293-7.293a1 1 0 011.414 0z"
-        clipRule="evenodd"
-      />
-    </svg>
-  );
-}
 
-/** X icon SVG */
-function XIcon() {
-  return (
-    <svg
-      xmlns="http://www.w3.org/2000/svg"
-      viewBox="0 0 20 20"
-      fill="currentColor"
-      style={{ width: "18px", height: "18px" }}
-      aria-hidden="true"
-    >
-      <path
-        fillRule="evenodd"
-        d="M4.293 4.293a1 1 0 011.414 0L10 8.586l4.293-4.293a1 1 0 111.414 1.414L11.414 10l4.293 4.293a1 1 0 01-1.414 1.414L10 11.414l-4.293 4.293a1 1 0 01-1.414-1.414L8.586 10 4.293 5.707a1 1 0 010-1.414z"
-        clipRule="evenodd"
-      />
-    </svg>
-  );
-}
 
 function ContrastBadge({
   pass,
@@ -557,25 +522,7 @@ export function ToolCanvas({
             style={{ alignSelf: "flex-start", marginTop: "0.25rem" }}
             aria-label={`Save current combination ${localFg} on ${localBg} (ratio ${formatRatio(ratio)})`}
           >
-            <svg
-              xmlns="http://www.w3.org/2000/svg"
-              viewBox="0 0 20 20"
-              fill="currentColor"
-              style={{
-                width: "16px",
-                height: "16px",
-                display: "inline-block",
-                verticalAlign: "middle",
-                marginRight: "0.25rem",
-              }}
-              aria-hidden="true"
-            >
-              <path
-                fillRule="evenodd"
-                d="M10 3a1 1 0 011 1v5h5a1 1 0 110 2h-5v5a1 1 0 11-2 0v-5H4a1 1 0 110-2h5V4a1 1 0 011-1z"
-                clipRule="evenodd"
-              />
-            </svg>
+            <PlusIcon />
             Save Combination
           </button>
         )}